optimize functions on contents_update table

+++ main/lib/idds/orm/contents.py
@@ -23,7 +23,7 @@ from sqlalchemy.sql.expression import asc
 
 from idds.common import exceptions
 from idds.common.constants import (ContentType, ContentStatus, ContentLocking,
-                                   ContentRelationType)
+                                   ContentFetchStatus, ContentRelationType)
 from idds.orm.base.session import read_session, transactional_session
 from idds.orm.base import models
 
@@ -560,7 +560,7 @@ def get_update_contents_from_others_by_dep_id(request_id=None, transform_id=None
 
 
 @read_session
-def get_updated_transforms_by_content_status(request_id=None, transform_id=None, session=None):
+def get_updated_transforms_by_content_status(request_id=None, transform_id=None, check_substatus=False, session=None):
     """
     Get updated transform ids by content status
 
@@ -588,9 +588,11 @@ def get_updated_transforms_by_content_status(request_id=None, transform_id=None,
         if request_id:
             query = query.filter(models.Content.request_id == request_id)
         query = query.filter(models.Content.content_relation_type == 3)
-        query = query.join(subquery, and_(models.Content.content_dep_id == subquery.c.content_id,
-                                          models.Content.substatus != subquery.c.substatus))
-
+        if check_substatus:
+            query = query.join(subquery, and_(models.Content.content_dep_id == subquery.c.content_id,
+                                              models.Content.substatus != subquery.c.substatus))
+        else:
+            query = query.join(subquery, and_(models.Content.content_dep_id == subquery.c.content_id))
         tmp = query.distinct()
 
         rets = []
@@ -685,19 +687,49 @@ def add_contents_update(contents, bulk_size=10000, session=None):
         raise exceptions.DatabaseException(error)
 
 
+@transactional_session
+def set_fetching_contents_update(request_id=None, transform_id=None, fetch=True, session=None):
+    """
+    Set fetching contents update.
+
+    :param session: session.
+    """
+    try:
+        if fetch:
+            query = session.query(models.Content_update)
+            if request_id:
+                query = query.filter(models.Content_update.request_id == request_id)
+            if transform_id:
+                query = query.filter(models.Content_update.transform_id == transform_id)
+            query.update({'fetch_status': ContentFetchStatus.Fetching})
+    except sqlalchemy.orm.exc.NoResultFound as error:
+        raise exceptions.NoObject('No record can be found with (transform_id=%s): %s' %
+                                  (transform_id, error))
+    except Exception as error:
+        raise error
+
+
 @read_session
-def get_contents_update(request_id=None, transform_id=None, session=None):
+def get_contents_update(request_id=None, transform_id=None, fetch=False, session=None):
     """
     Get contents update.
 
     :param session: session.
     """
     try:
-        query = session.query(models.Content_update)
-        if request_id:
-            query = query.filter(models.Content_ext.request_id == request_id)
-        if transform_id:
-            query = query.filter(models.Content_ext.transform_id == transform_id)
+        if fetch:
+            query = session.query(models.Content_update)
+            if request_id:
+                query = query.filter(models.Content_update.request_id == request_id)
+            if transform_id:
+                query = query.filter(models.Content_update.transform_id == transform_id)
+            query = query.filter(models.Content_update.fetch_status == ContentFetchStatus.Fetching)
+        else:
+            query = session.query(models.Content_update)
+            if request_id:
+                query = query.filter(models.Content_update.request_id == request_id)
+            if transform_id:
+                query = query.filter(models.Content_update.transform_id == transform_id)
 
         tmp = query.all()
         rets = []
@@ -713,7 +745,7 @@ def get_contents_update(request_id=None, transform_id=None, session=None):
 
 
 @transactional_session
-def delete_contents_update(request_id=None, transform_id=None, contents=[], bulk_size=10000, session=None):
+def delete_contents_update(request_id=None, transform_id=None, contents=[], bulk_size=1000, fetch=False, session=None):
     """
     delete a content.
 
@@ -723,27 +755,36 @@ def delete_contents_update(request_id=None, transform_id=None, contents=[], bulk
     :raises DatabaseException: If there is a database error.
     """
     try:
-        if contents:
-            contents_sub_params = [contents[i:i + bulk_size] for i in range(0, len(contents), bulk_size)]
-
-            for contents_sub_param in contents_sub_params:
-                del_query = session.query(models.Content_update)
-                if request_id:
-                    del_query = del_query.filter(models.Content_update.request_id == request_id)
-                if transform_id:
-                    del_query = del_query.filter(models.Content_update.transform_id == transform_id)
-                if contents_sub_param:
-                    del_query = del_query.filter(models.Content_update.content_id.in_(contents_sub_param))
-                del_query.with_for_update(nowait=True, skip_locked=True)
-                del_query.delete()
-        else:
+        if fetch:
             del_query = session.query(models.Content_update)
             if request_id:
                 del_query = del_query.filter(models.Content_update.request_id == request_id)
             if transform_id:
                 del_query = del_query.filter(models.Content_update.transform_id == transform_id)
-            del_query.with_for_update(nowait=True, skip_locked=True)
+            del_query = del_query.filter(models.Content_update.fetch_status == ContentFetchStatus.Fetching)
             del_query.delete()
+        else:
+            if contents:
+                contents_sub_params = [contents[i:i + bulk_size] for i in range(0, len(contents), bulk_size)]
+
+                for contents_sub_param in contents_sub_params:
+                    del_query = session.query(models.Content_update)
+                    if request_id:
+                        del_query = del_query.filter(models.Content_update.request_id == request_id)
+                    if transform_id:
+                        del_query = del_query.filter(models.Content_update.transform_id == transform_id)
+                    if contents_sub_param:
+                        del_query = del_query.filter(models.Content_update.content_id.in_(contents_sub_param))
+                    del_query.with_for_update(nowait=True, skip_locked=True)
+                    del_query.delete()
+            else:
+                del_query = session.query(models.Content_update)
+                if request_id:
+                    del_query = del_query.filter(models.Content_update.request_id == request_id)
+                if transform_id:
+                    del_query = del_query.filter(models.Content_update.transform_id == transform_id)
+                del_query.with_for_update(nowait=True, skip_locked=True)
+                del_query.delete()
     except Exception as error:
         raise exceptions.NoObject('Content_update deletion error: %s' % (error))
